% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calACSstr.R
\name{calACSstr}
\alias{calACSstr}
\title{Calculate the total number of all common subsequences between 2 strings}
\usage{
calACSstr(strA, strB, sep = "-", dropFirstItem = FALSE)
}
\arguments{
\item{strA}{First string}

\item{strB}{Second string}

\item{sep}{Delimiter separating each items in a sequence}

\item{dropFirstItem}{Boolean. If true, the first item in each sequence is excluded from counting all subsequences}
}
\value{
The total number of all common subsequences as an integer
}
\description{
Calculate the total number of all common subsequences between 2 strings
}
\examples{
calACSstr("q-w-e-r", "q-e-w-r", "-")
calACSstr("itemToBeDropped-q-w-e-r", "itemToBeDroped-q-e-w-r", "-", dropFirstItem=TRUE)

}

