% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePossibleSubsequences.r
\name{GeneratePossibleSubsequences}
\alias{GeneratePossibleSubsequences}
\title{Genearte all possible subsequences of the input vector. Require all items in the vector to be unique}
\usage{
GeneratePossibleSubsequences(vin)
}
\arguments{
\item{vin}{input vector that contains unique items}
}
\value{
a list of vectors, each a unique subsequence of the input vector
}
\description{
Genearte all possible subsequences of the input vector. Require all items in the vector to be unique
}

