% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.subvector.r
\name{is.subvector}
\alias{is.subvector}
\title{Check if the input vector subvec is a subvector of the other input vector vec}
\usage{
is.subvector(subvec, vec)
}
\arguments{
\item{subvec}{the input subvector}

\item{vec}{the input vector that may contain the subvector}
}
\description{
Check if the input vector subvec is a subvector of the other input vector vec
}

