% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenACSstrStrict.r
\name{lenACSstrStrict}
\alias{lenACSstrStrict}
\title{Internal function:
Calculate the length of the common subsequences between 2 input strings}
\usage{
lenACSstrStrict(strA, strB, sep = "-", dropFirstItem = FALSE,
  ignoreLenOneSubseq = FALSE)
}
\arguments{
\item{strA}{First string}

\item{strB}{Second string}

\item{sep}{Delimiter separating each items in a sequence}

\item{dropFirstItem}{Boolean. If true, the first item in each sequence is excluded from counting all subsequences}

\item{ignoreLenOneSubseq}{Boolean. If true, all length one subequences are not counted as common subsequences}
}
\value{
a vector containing the length of each common subsequence
}
\description{
Internal function:
Calculate the length of the common subsequences between 2 input strings
}

