% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenLCSstrStrict.r
\name{lenLCSstrStrict}
\alias{lenLCSstrStrict}
\title{Internal function:
Calculate the length of the longest common subsequence between 2 input strings}
\usage{
lenLCSstrStrict(strA, strB, sep = "-", dropFirstItem = FALSE)
}
\arguments{
\item{strA}{First string}

\item{strB}{Second string}

\item{sep}{Delimiter separating each items in a sequence}

\item{dropFirstItem}{Boolean. If true, the first item in each sequence is excluded from counting all subsequences}
}
\value{
a vector containing the length of each common subsequence
}
\description{
Internal function:
Calculate the length of the longest common subsequence between 2 input strings
}

