% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longestVec.r
\name{longestVec}
\alias{longestVec}
\title{The function takes in multiple vectors of any length, and returns the one with the longest length.
The tieBreaker variable controls if the first or the last of the longest vectors gets returned in case
there are multiple}
\usage{
longestVec(..., tieBreaker = "last")
}
\arguments{
\item{...}{vectors of any length}

\item{tieBreaker}{decides if the first or the last longest vector gets returned if there are multiple longest vectors.
Can be either 'first' or 'last'. Default to 'last'.}
}
\description{
The function takes in multiple vectors of any length, and returns the one with the longest length.
The tieBreaker variable controls if the first or the last of the longest vectors gets returned in case
there are multiple
}
\examples{
longestVec(1:5, c('a','b'))

}

