% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePossibleSubsequences.r
\name{GeneratePossibleSubsequences}
\alias{GeneratePossibleSubsequences}
\title{Internal function:
Genearte all possible subsequences of the input vector. Require all items in the vector to be unique}
\usage{
GeneratePossibleSubsequences(vin, ignoreLenOneSubseq = FALSE)
}
\arguments{
\item{vin}{input vector that contains unique items}

\item{ignoreLenOneSubseq}{Boolean. If true, all length one subequences are not counted as common subsequences}
}
\value{
a list of vectors, each a unique subsequence of the input vector
}
\description{
Internal function:
Genearte all possible subsequences of the input vector. Require all items in the vector to be unique
}

